requirements.txt
gdown
scikit-learn
numpy
pandas
python-dotenv
streamlit
